chore: move `docker_run_subprocess_step` to `crate::near::docker_build::subprocess_step`
