+++ cargo-near-build/Cargo.toml
@@ -28,11 +28,14 @@ pathdiff = { version = "0.2.1", features = ["camino"], optional = true }
 unix_path = { version = "1.0.1", optional = true }
 tempfile = { version = "3.10.1", optional = true }
 
+[target.'cfg(target_os = "linux")'.dependencies]
+nix = { version = "0.29.0", features = ["user", "process"], optional = true }
+
 [features]
 default = []
 docker = [
   "dep:url", "dep:serde", "dep:git2", 
   "dep:home", "dep:pathdiff", "dep:unix_path", 
-  "dep:tempfile"
+  "dep:tempfile", "dep:nix"
 ]
