+++ cargo-near-build/src/near/docker_build/mod.rs
@@ -1,2 +1,5 @@
 pub mod docker_checks;
 pub mod git_checks;
+pub mod subprocess_step;
+
+pub const ERR_REPRODUCIBLE: &str = "Reproducible build in docker container failed.";
