+++ cargo-near-build/src/types/near/docker_build/subprocess/env_vars/nep330_build_info.rs
