+++ cargo-near/Cargo.toml
@@ -49,8 +49,6 @@ tokio = { version = "1.0", default-features = false, features = [
 [dev-dependencies]
 tracing-test = "0.2.5"
 
-[target.'cfg(target_os = "linux")'.dependencies]
-nix = { version = "0.29.0", features = ["user", "process"] }
 
 [features]
 default = ["ledger"]
